third_party/blink/web_tests/platform/linux-chrome/external/wpt/picture-in-picture/picture-in-picture-window-expected.txt

This is a testharness.js-based test.
[FAIL] Picture-in-Picture window dimensions are set after entering Picture-in-Picture
  promise_test: Unhandled rejection with value: object "NotSupportedError: Picture-in-Picture is not available."
[FAIL] Picture-in-Picture window dimensions are set to 0 after entering Picture-in-Picture for another video
  promise_test: Unhandled rejection with value: object "NotSupportedError: Picture-in-Picture is not available."
[FAIL] Picture-in-Picture window is unchanged after entering Picture-in-Picture for video already in Picture-in-Picture
  promise_test: Unhandled rejection with value: object "NotSupportedError: Picture-in-Picture is not available."
[FAIL] Picture-in-Picture window dimensions are set to 0 after exiting Picture-in-Picture
  promise_test: Unhandled rejection with value: object "NotSupportedError: Picture-in-Picture is not available."
[FAIL] Picture-in-Picture window dimensions are set to 0 if disablePictureInPicture becomes true
  promise_test: Unhandled rejection with value: object "NotSupportedError: Picture-in-Picture is not available."
Harness: the test ran to completion.
